requirements.txt
astroid==1.4.8
awesome-slugify==1.6.5
blinker==1.4
click==6.6
flake8==3.0.3
Flask==0.11.1
Flask-Mail==0.9.1
flask-peewee==0.6.7
Frozen-Flask==0.12
gunicorn==19.6.0
isort==4.2.5
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.2.2
Markdown==2.6.6
MarkupSafe==0.23
mccabe==0.5.1
peewee==2.8.2
pycodestyle==2.0.0
pyflakes==1.2.3
pylint==1.6.4
regex==2016.7.21
requests==2.10.0
six==1.10.0
slugify==0.0.1
Unidecode==0.4.19
Werkzeug==0.11.10
wrapt==1.10.8
wtf-peewee==0.2.6
WTForms==2.1
pyyaml
